refactor: update imports to use models from src.config.models
