+++ src/services/utilities/__init__.py
@@ -1,6 +1,6 @@
 """Utility services and helpers."""
 
-from src.config import SearchAccuracy, VectorType
+from src.config.models import SearchAccuracy, VectorType
 from src.models.vector_search import PrefetchConfig, SearchStage
 
 from .hnsw_optimizer import HNSWOptimizer
