+++ tests/integration/browser/test_lightweight_scraper_integration.py
@@ -8,7 +8,8 @@ from pathlib import Path
 
 import pytest
 
-from src.config import Config, Environment
+from src.config import Config
+from src.config.models import Environment
 from src.services.browser.lightweight_scraper import LightweightScraper
