+++ tests/unit/services/browser/test_lightweight_scraper.py
@@ -8,7 +8,8 @@ import httpx
 import pytest
 import respx
 
-from src.config import Config, Environment
+from src.config import Config
+from src.config.models import Environment
 from src.services.browser.lightweight_scraper import LightweightScraper
